remove all s3.go and s3.go，move logic to all.go
